Move some related tests to a separate file
And fix/ignore offenses in the new file. This change is because this is
the biggest file we have and adding stuff to it (which is something very
frequent) makes RuboCop fail.
