+++ .rubocop_todo.yml
@@ -1,6 +1,6 @@
 # This configuration was generated by
 # `rubocop --auto-gen-config`
-# on 2017-04-02 09:58:55 -0300 using RuboCop version 0.46.0.
+# on 2017-04-02 12:28:33 -0300 using RuboCop version 0.46.0.
 # The point is for the user to remove these configuration records
 # one by one as the offenses are removed from the code base.
 # Note that changes in the inspected code, or installation of new
@@ -156,7 +156,7 @@ Lint/UnusedMethodArgument:
     - 'config/initializers/paperclip.rb'
     - 'test/support/base.rb'
 
-# Offense count: 14
+# Offense count: 13
 Lint/UselessAssignment:
   Exclude:
     - 'app/admin/collaboration.rb'
@@ -172,11 +172,6 @@ Lint/UselessAssignment:
     - 'test/models/collaboration_test.rb'
     - 'test/models/user_test.rb'
 
-# Offense count: 1
-Lint/Void:
-  Exclude:
-    - 'test/helpers/application_helper_test.rb'
-
 # Offense count: 87
 Metrics/AbcSize:
   Max: 149
@@ -190,16 +185,16 @@ Metrics/BlockLength:
 Metrics/BlockNesting:
   Max: 5
 
-# Offense count: 21
+# Offense count: 22
 # Configuration parameters: CountComments.
 Metrics/ClassLength:
-  Max: 692
+  Max: 602
 
 # Offense count: 37
 Metrics/CyclomaticComplexity:
   Max: 53
 
-# Offense count: 1359
+# Offense count: 1360
 # Configuration parameters: AllowHeredoc, AllowURI, URISchemes, IgnoreCopDirectives, IgnoredPatterns.
 # URISchemes: http, https
 Metrics/LineLength:
@@ -347,7 +342,7 @@ Style/AlignHash:
     - 'app/models/impulsa_edition.rb'
     - 'lib/podemos_import_collaborations.rb'
 
-# Offense count: 17
+# Offense count: 14
 # Cop supports --auto-correct.
 # Configuration parameters: EnforcedStyle, SupportedStyles, IndentationWidth.
 # SupportedStyles: with_first_parameter, with_fixed_indentation
@@ -363,7 +358,6 @@ Style/AlignParameters:
     - 'bin/rubocop'
     - 'config/initializers/amazon_ses.rb'
     - 'test/integration/verification_online_manager_test.rb'
-    - 'test/models/user_test.rb'
 
 # Offense count: 188
 # Cop supports --auto-correct.
@@ -485,7 +479,7 @@ Style/CommentAnnotation:
     - 'test/controllers/vote_controller_test.rb'
     - 'test/integration/proposals_test.rb'
 
-# Offense count: 24
+# Offense count: 22
 # Cop supports --auto-correct.
 Style/CommentIndentation:
   Exclude:
@@ -496,7 +490,6 @@ Style/CommentIndentation:
     - 'config/application.rb'
     - 'test/controllers/collaborations_controller_test.rb'
     - 'test/models/order_test.rb'
-    - 'test/models/user_test.rb'
 
 # Offense count: 11
 # Cop supports --auto-correct.
@@ -651,7 +644,7 @@ Style/FormatString:
     - 'test/models/collaboration_test.rb'
     - 'test/models/microcredit_test.rb'
 
-# Offense count: 246
+# Offense count: 247
 # Cop supports --auto-correct.
 # Configuration parameters: EnforcedStyle, SupportedStyles.
 # SupportedStyles: when_needed, always
@@ -721,7 +714,7 @@ Style/IndentHash:
     - 'test/models/collaboration_test.rb'
     - 'test/models/microcredit_loan_test.rb'
 
-# Offense count: 9
+# Offense count: 8
 # Cop supports --auto-correct.
 # Configuration parameters: EnforcedStyle, SupportedStyles.
 # SupportedStyles: normal, rails
@@ -732,7 +725,6 @@ Style/IndentationConsistency:
     - 'app/models/election.rb'
     - 'app/models/user.rb'
     - 'lib/tasks/encomu/geoextra_extract.rake'
-    - 'test/helpers/application_helper_test.rb'
     - 'test/models/collaboration_test.rb'
 
 # Offense count: 19
@@ -1209,7 +1201,7 @@ Style/SpaceInLambdaLiteral:
 # Cop supports --auto-correct.
 Style/SpaceInsideArrayPercentLiteral:
   Exclude:
-    - 'test/models/user_test.rb'
+    - 'test/models/concerns/verificable_test.rb'
 
 # Offense count: 225
 # Cop supports --auto-correct.
@@ -1273,7 +1265,7 @@ Style/StabbyLambdaParentheses:
   Exclude:
     - 'app/models/order.rb'
 
-# Offense count: 1606
+# Offense count: 1613
 # Cop supports --auto-correct.
 # Configuration parameters: EnforcedStyle, SupportedStyles, ConsistentQuotesInMultiline.
 # SupportedStyles: single_quotes, double_quotes
@@ -1347,7 +1339,7 @@ Style/TrailingCommaInLiteral:
     - 'app/controllers/open_id_controller.rb'
     - 'config/initializers/geoextra.rb'
 
-# Offense count: 427
+# Offense count: 426
 # Cop supports --auto-correct.
 Style/TrailingWhitespace:
   Enabled: false
@@ -1380,7 +1372,7 @@ Style/VariableName:
   Exclude:
     - 'lib/tasks/encomu/municipies_extract.rake'
 
-# Offense count: 135
+# Offense count: 133
 # Cop supports --auto-correct.
 # Configuration parameters: SupportedStyles, WordRegex.
 # SupportedStyles: percent, brackets
